apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
psycopg2-binary>=2.9.10
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
python-multipart>=0.0.18
redis>=5.2.0
httpx>=0.28.0
email-validator>=2.2.0
